Reworked the comment entities
