+++ src/main/java/nl/tudelft/ewi/devhub/server/web/errors/EntityNotFoundExceptionMapper.java
@@ -42,7 +42,9 @@ public class EntityNotFoundExceptionMapper implements ExceptionMapper<EntityNotF
 
 		try {
 			Map<String, Object> params = Maps.newHashMap();
-			params.put("user", determineUser());
+			User user = determineUser();
+			if(user != null)
+				params.put("user", determineUser());
 			params.put("error_id", id);
 
 			return Response
