+++ src/main/resources/static/less/devhub.less
@@ -723,6 +723,13 @@ code {
   margin-bottom: 10px !important;
 }
 
+.panel-comment {
+  p {
+	white-space: pre;
+	margin: 0;
+  }
+}
+
 .pull-feed {
   position: relative;
   padding: 10px 0;
